# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dataplat API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/operators": {
      "get": {
        "tags": [
          "operators"
        ],
        "summary": "List Operators",
        "description": "List active operators, optionally filtered by category.\n\nReturns a plain JSON array of OperatorRead objects. All authenticated users\ncan list operators \u2014 no ownership scoping applies (operators are global\nregistry entries, not user-owned resources).\n\nActive scoping:\n  `is_active IS NOT FALSE` \u2014 rows with is_active=true OR is_active=NULL are\n  included; rows with is_active=false are excluded. This is the correct\n  expression when the column is nullable with a server-side default of true.\n\nCategory filter:\n  If ?category= is provided, further restricts to operators with that exact\n  category value. An unknown category value returns HTTP 200 + [].\n\nOrdering: id ASC (stable, oldest-first, consistent with all other list endpoints).\n\nAuth required (F-008).",
        "operationId": "list_operators_api_operators_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by operator category (e.g. 'extractor', 'tagger'). Optional. Unknown categories return an empty array.",
              "title": "Category"
            },
            "description": "Filter by operator category (e.g. 'extractor', 'tagger'). Optional. Unknown categories return an empty array."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperatorRead"
                  },
                  "title": "Response List Operators Api Operators Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/operators/{operator_id}": {
      "get": {
        "tags": [
          "operators"
        ],
        "summary": "Get Operator Detail",
        "description": "Return the full operator record for the given id.\n\nOperators are a global registry \u2014 there is no owner_id column and no\nownership scoping. Any authenticated user can retrieve any operator by id.\nThis differs from sources (owner-scoped) and collections (owner-scoped).\n\nReturns 404 if no operator row with operator_id exists.\nReturns 401 if the Bearer token is absent, invalid, or expired.\nReturns 422 if operator_id is not a valid integer (FastAPI default).\n\nAuth required (F-008).",
        "operationId": "get_operator_api_operators__operator_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operator_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Operator Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperatorDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "input_kind": {
            "type": "string",
            "title": "Input Kind"
          },
          "output_kind": {
            "type": "string",
            "title": "Output Kind"
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "output_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Schema"
          },
          "config_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Schema"
          },
          "default_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Config"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reference_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Url"
          },
          "example_input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Input"
          },
          "example_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example Output"
          },
          "entrypoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrypoint"
          },
          "estimated_cost_per_unit": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Per Unit"
          },
          "rate_limit_per_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Minute"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "category",
          "input_kind",
          "output_kind",
          "image",
          "output_schema",
          "config_schema",
          "default_config",
          "description",
          "reference_url",
          "example_input",
          "example_output",
          "entrypoint",
          "estimated_cost_per_unit",
          "rate_limit_per_minute",
          "is_active",
          "created_at"
        ],
        "title": "OperatorDetail",
        "description": "Full operator record for GET /api/operators/{operator_id} (F-017).\n\nExposes all 19 ORM columns. Nullability matches Mapped[Optional[...]] exactly:\n  NOT NULL: id, name, version, category, input_kind, output_kind, image.\n  nullable:  output_schema, config_schema, default_config, description,\n             reference_url, example_input, example_output, entrypoint,\n             estimated_cost_per_unit, rate_limit_per_minute, is_active, created_at.\n\ndefault_config has server_default '{}'::jsonb \u2014 a fresh SELECT always returns\na dict (empty is valid). output_schema is NULL for the MinerU seed row (seed\nnever sets it); the key is present in the response but the value is None."
      },
      "OperatorRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "input_kind": {
            "type": "string",
            "title": "Input Kind"
          },
          "output_kind": {
            "type": "string",
            "title": "Output Kind"
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "config_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Schema"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "category",
          "input_kind",
          "output_kind",
          "image",
          "config_schema",
          "description",
          "is_active"
        ],
        "title": "OperatorRead",
        "description": "Response schema for a single operator row.\n\nUsed by GET /api/operators (F-016).\n\nFields included per agreed.md \u00a73.4:\n  - id, name, version, category \u2014 required by F-016 V2.\n  - config_schema \u2014 required by F-016 V2; nullable (Optional[dict] in ORM).\n  - input_kind, output_kind \u2014 useful to callers composing pipelines.\n  - image \u2014 operators are identified by their container image.\n  - description \u2014 human-readable label for UI display; nullable.\n  - is_active \u2014 lets clients reason about operator availability; nullable bool.\n\nFields intentionally omitted (internal/operational details deferred to later\nsprints): output_schema, default_config, reference_url, example_input,\nexample_output, entrypoint, estimated_cost_per_unit, rate_limit_per_minute,\ncreated_at."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}